Revert "Closing in on end"
This reverts commit 477a36ed8a32f122804c79d669feff812368231e.

+++ TravelPlanner/TravelPlannerApp/Repository/Database/MockCapitalDatabase.cs
@@ -7,7 +7,7 @@ namespace TravelPlanner.TravelPlannerApp.Repository.Database
     internal class MockCapitalDatabase : IMockDatabase
     {
         private readonly List<Capital> _capitalList = new();
-        private int _mockCurrentId = 0;
+        private long _mockCurrentId = 0;
 
         public MockCapitalDatabase()
         {
